+++ sample/Item.cs
@@ -1,11 +1,10 @@
 ﻿using System.Text.Json.Serialization;
+using Tavenem.DataStorage;
 
 namespace Tavenem.Blazor.IndexedDB.Sample;
 
-public class Item
+public class Item : IdItem
 {
-    public string? Id { get; set; }
-
     [JsonIgnore] public bool IsUpdating { get; set; }
 
     [JsonIgnore] public string? NewValue { get; set; }