+++ sample/Program.cs
@@ -10,6 +10,6 @@ builder.RootComponents.Add<HeadOutlet>("head::after");
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 
 builder.Services.AddIndexedDb(
-    new IndexedDb<string>("Tavenem.Blazor.IndexedDB.Sample", 1));
+    new IndexedDb("Tavenem.Blazor.IndexedDB.Sample", 1));
 
 await builder.Build().RunAsync().ConfigureAwait(false);