+++ src/assets/tavenem-indexeddb.js
@@ -59,7 +59,7 @@ export async function deleteValue(databaseInfo, key) {
         return false;
     }
     try {
-        return await db.delete(databaseInfo.storeName ?? databaseInfo.databaseName, key);
+        await db.delete(databaseInfo.storeName ?? databaseInfo.databaseName, key);
         return true;
     }
     catch (e) {