+++ docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0-preview1
+### Changed
+- `IndexedDb` no longer has a typed key. Only string keys are used.
+- `IndexedDbService` no longer has a typed key. Only string keys are used.
+- `IndexedDbService` now implements [`IDataStore`](https://github.com/Tavenem/DataStore).
+  - `AddValueAsync` and `PutValueAsync` are now `StoreItemAsync<T>`
+  - `GetValueAsync` is now `GetItemAsync<T>`
+  - `DeleteValueAsync` is now `RemoveItemAsync<T>`
+
 ## 1.0.0
 ### Added
 - Initial release